Renames `immutable_buffer.rs` files to `binary_buffer.rs`
