+++ src/lazy/binary/mod.rs
@@ -1,5 +1,5 @@
+pub mod binary_buffer;
 mod encoded_value;
-pub mod immutable_buffer;
 pub mod raw;
 
 #[cfg(test)]
