+++ src/lazy/binary/raw/v1_1/mod.rs
@@ -1,6 +1,6 @@
 mod annotations_iterator;
 pub use annotations_iterator::*;
-pub mod immutable_buffer;
+pub mod binary_buffer;
 pub mod reader;
 pub mod sequence;
 pub mod r#struct;
