+++ src/lib.rs
@@ -314,7 +314,7 @@ macro_rules! v1_0_tooling_apis {
             binary::uint::DecodedUInt,
             binary::var_int::VarInt,
             binary::var_uint::VarUInt,
-            lazy::binary::immutable_buffer::{BinaryBuffer, AnnotationsWrapper},
+            lazy::binary::binary_buffer::{BinaryBuffer, AnnotationsWrapper},
             lazy::binary::raw::type_descriptor::Header,
             lazy::raw_value_ref::RawValueRef,
             lazy::encoder::binary::v1_0::writer::LazyRawBinaryWriter_1_0 as RawBinaryWriter,
